apps/hubspot/src/app/setup/service.test.ts
/**
 * DISCLAIMER:
 * The tests provided in this file are specifically designed for the `setupOrganisation` function.
 * These tests illustrate potential scenarios and methodologies relevant for SaaS integration.
 * Developers should create tests tailored to their specific implementation and requirements.
 * Mock data and assertions here are simplified and may not cover all real-world complexities.
 * Expanding upon these tests to fit the actual logic and behaviors of specific integrations is crucial.
 */
import { expect, test, describe, vi, beforeAll, afterAll } from 'vitest';
import { eq } from 'drizzle-orm';
import * as authConnector from '@/connectors/auth';
import * as accountDetailsConnector  from '@/connectors/account-info';
import { db } from '@/database/client';
import { Organisation } from '@/database/schema';
import { inngest } from '@/inngest/client';
import { setupOrganisation } from './service';

const code = 'some-code';
const accessToken = 'some token';
const expiresIn = 60;
const region = 'us';
const timeZone = 'us/eastern'
const now = new Date();
const getTokenData = {
  id:"45a76301-f1dd-4a77-b12f-9d7d3fca3c90",
  createdAt:"2024-02-07T04:37:25.229Z",
  accessToken,
  refreshToken: "some token",
  expiresIn,
};

const organisation = {
  id: '45a76301-f1dd-4a77-b12f-9d7d3fca3c90',
  accessToken: 'COfmhPjXMRIOSAEAQAAAAAAAAAAAAAEY1Mq5FSDnzpEeKN3rpwEyFKtO02yusybc9RD4E8fAJxGh87KBOjAAAAFHAAAAAAAAQP-AHwAAAIAAAAAAAAAAAAAAAAAAAABgAAAAACAAAAAAAAAQAAJCFG4ji356a7MskwieyoEeQm38d4b2SgNuYTFSAFoA',
  refreshToken:"some refreshtoken",
  region,
  timeZone
};

describe('setupOrganisation', () => {
  beforeAll(() => {
    vi.setSystemTime(now);
  });

  afterAll(() => {
    vi.useRealTimers();
  });

  test('should setup organisation when the code is valid and the organisation is not registered', async () => {
    // mock inngest client, only inngest.send should be used
    // @ts-expect-error -- this is a mock
    const send = vi.spyOn(inngest, 'send').mockResolvedValue(undefined);
    // mock the getToken function to return a predefined token
    const getToken = vi.spyOn(authConnector, 'getToken').mockResolvedValue(getTokenData);

    // mock the getAccountDetails function to return a predefined accountDetails
    const getAccountDetails = vi.spyOn(accountDetailsConnector, 'getAccountDetails').mockResolvedValue(timeZone);

    // assert the function resolves without returning a value
    await expect(
      setupOrganisation({
        organisationId: organisation.id,
        code,
        region,
      })
    ).resolves.toBeUndefined();

    // check if getToken was called correctly
    expect(getToken).toBeCalledTimes(1);
    expect(getToken).toBeCalledWith(code);

    // check if getToken was called correctly
    expect(getAccountDetails).toBeCalledTimes(1);
    expect(getAccountDetails).toBeCalledWith(accessToken);

    // verify the organisation token is set in the database
    await expect(
      db.select().from(Organisation).where(eq(Organisation.id, organisation.id))
    ).resolves.toMatchObject([
      {
        accessToken,
        region,
      },
    ]);

    // verify that the user/sync event is sent
    expect(send).toBeCalledTimes(1);
    expect(send).toBeCalledWith([
      {
        name: 'hubspot/users.sync.requested',
        data: {
          isFirstSync: true,
          organisationId: organisation.id,
          syncStartedAt: now.getTime(),
          region,
          page: "",
        },
      },
      {
        name: 'hubspot/hubspot.elba_app.installed',
        data: {
          organisationId: organisation.id,
          region,
        },
      },
      {
        name: 'hubspot/hubspot.token.refresh.requested',
        data: {
          organisationId: organisation.id,
          region,
        },
        ts: now.getTime() + (expiresIn - 5) * 60 * 1000,
      },
      {
        name: 'hubspot/timezone.refresh.requested',
        data: {
          organisationId: organisation.id,
          region,
        },
      },
    ]);
  });

  test('should setup organisation when the code is valid and the organisation is already registered', async () => {
    // mock inngest client, only inngest.send should be used
    // @ts-expect-error -- this is a mock
    const send = vi.spyOn(inngest, 'send').mockResolvedValue(undefined);
    // pre-insert an organisation to simulate an existing entry
    await db.insert(Organisation).values(organisation);

    // mock getToken as above
    const getToken = vi.spyOn(authConnector, 'getToken').mockResolvedValue(getTokenData);

    // mock the getAccountDetails function to return a predefined accountDetails
    const getAccountDetails = vi.spyOn(accountDetailsConnector, 'getAccountDetails').mockResolvedValue(timeZone);

    // assert the function resolves without returning a value
    await expect(
      setupOrganisation({
        organisationId: organisation.id,
        code,
        region,
      })
    ).resolves.toBeUndefined();

    // verify getToken usage
    expect(getToken).toBeCalledTimes(1);
    expect(getToken).toBeCalledWith(code);

    // verify getToken usage
    expect(getAccountDetails).toBeCalledTimes(1);
    expect(getAccountDetails).toBeCalledWith(accessToken);
    // check if the token in the database is updated
    await expect(
      db
        .select({ accessToken: Organisation.accessToken })
        .from(Organisation)
        .where(eq(Organisation.id, organisation.id))
    ).resolves.toMatchObject([
      {
        accessToken,
      },
    ]);

    // verify that the user/sync event is sent
    expect(send).toBeCalledTimes(1);
    expect(send).toBeCalledWith([
      {
        name: 'hubspot/users.sync.requested',
        data: {
          isFirstSync: true,
          organisationId: organisation.id,
          syncStartedAt: now.getTime(),
          region,
          page: "",
        },
      },
      {
        name: 'hubspot/hubspot.elba_app.installed',
        data: {
          organisationId: organisation.id,
          region,
        },
      },
      {
        name: 'hubspot/hubspot.token.refresh.requested',
        data: {
          organisationId: organisation.id,
          region,
        },
        ts: now.getTime() + (expiresIn - 5) * 60 * 1000,
      },
      {
        name: 'hubspot/timezone.refresh.requested',
        data: {
          organisationId: organisation.id,
          region,
        },
      },
    ]);
  });

  test('should not setup the organisation when the code is invalid', async () => {
    // mock inngest client
    // @ts-expect-error -- this is a mock
    const send = vi.spyOn(inngest, 'send').mockResolvedValue(undefined);
    const error = new Error('invalid code');
    // mock getToken to reject with a dumb error for an invalid code
    const getToken = vi.spyOn(authConnector, 'getToken').mockRejectedValue(error);

    // assert that the function throws the mocked error
    await expect(
      setupOrganisation({
        organisationId: organisation.id,
        code,
        region,
      })
    ).rejects.toThrowError(error);

    // verify getToken usage
    expect(getToken).toBeCalledTimes(1);
    expect(getToken).toBeCalledWith(code);

    // ensure no organisation is added or updated in the database
    await expect(
      db.select().from(Organisation).where(eq(Organisation.id, organisation.id))
    ).resolves.toHaveLength(0);

    // ensure no sync users event is sent
    expect(send).toBeCalledTimes(0);
  });
});
